Complete Router for CRUD and SMS Verify
Fix random number placement in the send_sms
Complete Mail Verification
Router:
	- Start: user.router.js -> user.controller.js -> user.service.js
