+++ back-end/myapp/package.json
@@ -24,7 +24,8 @@
     "morgan": "~1.9.1",
     "mysql": "^2.18.1",
     "nodemon": "^2.0.7",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "twilio": "^3.61.0"
   },
   "devDependencies": {
     "chai": "^4.3.4",
